Ground-up rewrite of spark-bench
This commit represents a ground-up rewrite of spark-bench.

- From bash and separate maven builds to single, unified
  Scala project build with SBT.
- workloads and data generators now part of extensible
  framework and can be easily added
- Extensive configuration capability with a Typesafe
  configuration file

Configurable features include
- three different levels of parallelism: spark-submit,
  workload suites, within one workload suite
- easy repetition of workloads
- input datasets from any Hadoop-friendly system
  (local, HDFS, S3, etc) in any Hadoop-friendly
  format (CSV, Parquet, etc)
- output results to any Hadoop-friendly system in a
  variety of Hadoop-friendly formats
- easily manage all exposed parameters of any workload

This rewrite is not feature-complete with mainline spark-bench.
Rather, it encompasses a few select workloads and data generators
pulled from mainline with a few new workloads to expand the capability
of users.
The full list of workloads available in this version of the rewrite is
available in docs/workloads-and-parameters.

Contributors to this work:

Emily Curtin
Matthew Schauer
Brad Kaiser
Craig Ingram

Closes #51