Specifying Precise to prevent build failures
This is a bandaid for an issue that will require a deeper fix in the future, and may require a fix from TravisCI.
Builds are failing in TravisCI Trusty containers that are passing in Precise containers. I have looked into permissions issues, file structures, etc. and there are no culprits in those directions.
The issue appears to be a memory issue in the Trusty containers.

This fix will allow builds to pass for now, but TravisCI is eventually EOL'ing their Precise support so a real fix for the issue, whether from our side or TravisCI's, should be implemented.

Further investigation as to the cause behind the failures will require TravisCI support.