Sleep workload can use distribution sampling
This refactors the Sleep workload so that users have a variety of
options for using it. Before, the Sleep workload could either accept a
number of milliseconds to sleep the thread, or it would generate a
number less than a given max.

Now, users still have the option of providing a simple number of
milliseconds as before, but they also have the option of drawing from a
uniform distribution with min and max, a Poisson distribution with a
given mean (especially useful for simulating multiple users on a
cluster), or a Gaussian distribution with a mean and standard deviation.
This model is extensible to other distributions in the future should
users require.

The distribution functionality is built using the Breeze library.