Adding graph data generator
This commit adds a graph data generator to the workloads. Using the
GraphX logNormalGraph generator, users can pass the number of verices
and the distribution parameters for the edges to generate a graph.

Due to limitations in the GraphX GraphLoader, this generator can only
write to .txt format.

Additionally this commit provides tests for the GraphDataGen and
appropriate documentation updates. Also, I added a badge to the README
to point to the release that will be generated on Github Releases.