Passing config string instead of temp file
When launching spark-bench through assembled spark-submit scripts, the
child argument used to be a path to a temporary file created by the
launch process. This was causing issues when running in cluster mode
because the temporary files were in local storage.

Now instead the configuration is passed as one long string instead of a
path to a temporary file.