Stripping README, improving docs site
This commit moves all project documentation from the README to the docs
site. This will hopefully be more user-friendly and prevent the README
documentation from getting stale.

This commit also adds a media page with the Spark-Bench presentation, a
guide for contributing to Spark-Bench, improvements to the Quickstart
guide, addition of an Installation guide, and a guide for using SBT in
the context of the project.